fix: #73 update load more when change field sort
